Move statistics folder
